+++ pfvs/conipfv_kernel/conipfv_kernel.pyx
@@ -24,7 +24,7 @@ cdef extern from "gmp.h":
 cdef extern from "conipfv_kernel.h":
     int _conipfv_kernel_c(
         int32_t *out,
-        float *Qs,
+        double *Qs,
         int *N_out,
         int dim,
         double *U,
@@ -146,7 +146,7 @@ def conipfv_kernel(U,
             mpz_clear(H_gmp[i])
         free(H_gmp)
         raise MemoryError("Failed to allocate c_out")
-    cdef float *c_Qs = <float *>malloc(max_N_out * sizeof(float))
+    cdef double *c_Qs = <double *>malloc(max_N_out * sizeof(double))
     if c_Qs == NULL:
         for i in range(dim * dim):
             mpz_clear(H_gmp[i])
